+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Added custom resource in the lca-websocket-stack that sets the ECS cluster's desired service count to zero, then deletes the service, before deleting the cluster. This will prevent errors in deleting the stack when disabling websockets or deleting LCA.
+### Added
+- Support for Amazon Connect Kinesis Video Streams as an audio ingestion source. See [Amazon Connect KVS README](./lca-connect-kvs-stack/README.md) for more details.
 
 ## [0.8.10] - 2023-11-25
 ### Added